翻译go接口三个

+++ longjian-measure-app/src/main/java/com/longfor/longjian/measure/app/vo/proMeasureVo/ProMeasurePlanVo.java
@@ -9,7 +9,7 @@ import lombok.Data;
 @Data
 public class ProMeasurePlanVo {
     private String close_status;
-    private Integer create_at;
+    private Long create_at;
     private String finish_status;
     private Integer id;
     private Integer issue_count;
